Move Auth0ClusterSDK service to new Auth0ServiceProvider
