+++ app/Providers/AppServiceProvider.php
@@ -4,12 +4,8 @@ declare(strict_types=1);
 
 namespace App\Providers;
 
-use App\Auth0\Auth0ClusterSDK;
-use App\Auth0\Auth0Tenant;
-use App\Auth0\Auth0TenantSDK;
 use App\Insightly\InsightlyClient;
 use App\Insightly\Pipelines;
-use Auth0\SDK\Configuration\SdkConfiguration;
 use GuzzleHttp\Client;
 use Illuminate\Support\ServiceProvider;
 
@@ -27,29 +23,6 @@ final class AppServiceProvider extends ServiceProvider
                 ),
                 config('insightly.api_key'),
                 new Pipelines(config('insightly.pipelines'))
-            );
-        });
-
-        $this->app->singleton(Auth0ClusterSDK::class, function () {
-            return new Auth0ClusterSDK(
-                ...array_map(
-                    static fn (array $tenantConfig, string $tenant) => new Auth0TenantSDK(
-                        Auth0Tenant::from($tenant),
-                        new SdkConfiguration(
-                            domain: $tenantConfig['domain'],
-                            clientId: $tenantConfig['clientId'],
-                            clientSecret: $tenantConfig['clientSecret'],
-                            audience: [$tenantConfig['audience']],
-                        )
-                    ),
-                    array_filter(
-                        config('auth0.tenants'),
-                        static fn (array $tenantConfig) =>
-                            $tenantConfig['domain'] !== '' &&
-                            $tenantConfig['clientId'] !== '' &&
-                            $tenantConfig['clientSecret'] !== ''
-                    )
-                )
             );
         });
     }
